Initial commit.
This software is a project of the Joint Video Team of the ISO/IEC MPEG and
the ITU-T VCEG, retrieved from a CVS repository. The changes I'm making are
to fix code that breaks compilation on my system (Linux Mint x86_64, gcc 4.7.3)
and to modify the project for the purposes of a research project at Portland
State University.